Call updateElement iff element is added
Because player.addElement is asynchronous, player.updateElement can be
called before the element is added to player.

In that case, player.updateElement just returns a `Promise.reject()`,
and it is recorded as an error of `Uncaught (in promise) undefined`
in web console.

This change makes VFXElement track addition of its element and
call player.updateElement only if element is added.